+++ extendr-api/src/robj/try_from_robj.rs
@@ -579,5 +579,63 @@ impl TryFrom<&Robj> for HashMap<&str, Robj> {
             .collect::<HashMap<&str, Robj>>())
     }
 }
+
+macro_rules! impl_try_from_robj_for_arrays {
+    ($slice_type:ty) => {
+        impl<const N: usize> TryFrom<&Robj> for [$slice_type; N] {
+            type Error = Error;
+
+            fn try_from(value: &Robj) -> Result<Self> {
+                let value: &[$slice_type] = value.try_into()?;
+                if value.len() != N {
+                    return Err(Error::ExpectedLength(N));
+                }
+                let value: Self = value
+                    .try_into()
+                    .map_err(|error| format!("{}", error).to_string())?;
+                Ok(value)
+            }
+        }
+
+        // TODO: the following can be integrated into `impl_try_from_robj` later
+
+        impl<const N: usize> TryFrom<Robj> for [$slice_type; N] {
+            type Error = Error;
+
+            fn try_from(robj: Robj) -> Result<Self> {
+                Self::try_from(&robj)
+            }
+        }
+
+        impl<const N: usize> TryFrom<&Robj> for Option<[$slice_type; N]> {
+            type Error = Error;
+
+            fn try_from(robj: &Robj) -> Result<Self> {
+                if robj.is_null() || robj.is_na() {
+                    Ok(None)
+                } else {
+                    Ok(Some(<[$slice_type; N]>::try_from(robj)?))
+                }
+            }
+        }
+
+        impl<const N: usize> TryFrom<Robj> for Option<[$slice_type; N]> {
+            type Error = Error;
+
+            fn try_from(robj: Robj) -> Result<Self> {
+                Self::try_from(&robj)
+            }
+        }
+    };
+}
+
+impl_try_from_robj_for_arrays!(Rint);
+impl_try_from_robj_for_arrays!(Rfloat);
+impl_try_from_robj_for_arrays!(Rbool);
+impl_try_from_robj_for_arrays!(Rcplx);
+impl_try_from_robj_for_arrays!(u8);
+impl_try_from_robj_for_arrays!(i32);
+impl_try_from_robj_for_arrays!(f64);
+
 // Choosing arity 12.. As the Rust compiler did for these [Tuple to array conversion](https://doc.rust-lang.org/stable/std/primitive.tuple.html#trait-implementations-1)
 impl_try_from_robj_tuples!((1, 12));