+++ tests/extendrtests/src/rust/src/tuple_conversions.rs
@@ -35,8 +35,50 @@ fn sum_points(x: (Point, Point)) -> Doubles {
     Doubles::from_values([x1 + x2, y1 + y2])
 }
 
+#[extendr]
+fn round_trip_array_u8(x: [u8; 4]) -> [u8; 4] {
+    x
+}
+
+#[extendr]
+fn round_trip_array_f64(x: [f64; 4]) -> [f64; 4] {
+    x
+}
+
+#[extendr]
+fn round_trip_array_i32(x: [i32; 4]) -> [i32; 4] {
+    x
+}
+
+#[extendr]
+fn round_trip_array_rint(x: [Rint; 4]) -> [Rint; 4] {
+    x
+}
+
+#[extendr]
+fn round_trip_array_rfloat(x: [Rfloat; 4]) -> [Rfloat; 4] {
+    x
+}
+
+#[extendr]
+fn round_trip_array_rbool(x: [Rbool; 4]) -> [Rbool; 4] {
+    x
+}
+
+#[extendr]
+fn round_trip_array_rcplx(x: [Rcplx; 4]) -> [Rcplx; 4] {
+    x
+}
+
 extendr_module! {
     mod tuple_conversions;
     fn sum_triplet_ints;
     fn sum_points;
+    fn round_trip_array_f64;
+    fn round_trip_array_i32;
+    fn round_trip_array_rbool;
+    fn round_trip_array_rcplx;
+    fn round_trip_array_rfloat;
+    fn round_trip_array_rint;
+    fn round_trip_array_u8;
 }